feat: переместить REST контроллеры в package api
